% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_script_interactive.R
\name{submit_script_interactive}
\alias{submit_script_interactive}
\title{Submit an R Script}
\usage{
submit_script_interactive()
}
\description{
Transform an R script into a notebook task, installs R packages if needed,
and submits as job to a Databricks cluster.
}
